implemented a method to allow for an untimed waiting for reserving a client

+++ inc/client_set.h
@@ -51,8 +51,9 @@ unsigned int get_max_clients(client_set* cls);
 // returns 0 on failure, it fails only if shutdown was called earlier
 int reset_max_clients(client_set* cls, unsigned int max_clients);
 
-// this function returns NULL, only when the shutdown is called for the client_set
-stream* reserve_client(client_set* cls, unsigned long long int timeout_in_secs);
+// this function returns NULL, only when the shutdown is called for the client_set OR if you timed out
+// pass timeout as 0, to perform an untimed wait
+stream* reserve_client(client_set* cls, unsigned int timeout_in_secs);
 
 // return a reserved client, this function will close and destroy the client stream if it has errored
 void return_client(client_set* cls, stream* strm);
